Adding a helper message that fixes the date string.
I don't need this change right now, given the fact that I am displaying the date as a string in both cases (on the Achievement Table and on the Achievement Details page). However when the edit case gets implemented, I will need the date to be displayed correctly in that input.